backend: don't fake the list of tests
This is an ugly testament from the time of rash re-implementations of
all the backend logic.  Many tests were disabled, and made hidden
without making them work again - and we lost the track so we broke them
even more later.  Let's stop doing that now, make all the tests visible
but skipped (except for test_job_grab.py, which seems to be entirely
useless) and make them work soon.

This is a reaction on nice review from PR#1007.