backend: lower the traffic in build_dispatcher.log
With large pending build queue, some of the messages got printed for
each pending task each time the task was fetched from frontend.  This
was even tens of tousands of messages each 'sleeptime' (30 seconds).

Change level of repeatedly written messages from INFO to DEBUG, so we
don't spam the log (ATM it has ~12GB in production).  When needed, we
opt-in the DEBUG output with 'log_level = debug' config option.