backend: get_redis_logger: skip log entries bellow log_level
The RedisLogHandler already reads and respects log_level option from
configuration (by default INFO), so we don't actually waste the space in
log files.  But it is even useless to overwhelm redis server with
messages which are eventually not going to be printed to log file.  So
from now, the "level" is respected also on log entry "sender" side (and
I personally think that we could respect the level _only_ on sender
side, and let the RedisLogHandler dump everything is received, ... maybe
next PR?).

This is reaction for magically disappearing log entries upon Worker
failure (exceptions should be logged, but are not).  I suppose that the
redis server isn't ready to process all the log entries, and it leads to
obscure system states.

Merges: #1065