docker: set backend URL to a hostname known by other containers
The problem is following - http://localhost:5002 is approachable
from the host system and backend results are served there. This URL
needs to be used e.g. when rendering links pointing to logs on
the backend. However, at the same time, this URL is not available
within the containers so builds fail when using it for `copr_base`
repository with `skip_if_unavailable=False`.

I don't know how to get both features working other than having
two distinct config options for `BACKEND_BASE_URL` (e.g. adding
`BACKEND_BASE_URL_PRIVATE`). And I am not a big fan of this solution.