backend: run playbooks with timeout
In some circumstances, playbook spawning can indefinitely stuck on something
and therefore block the builder VM without a possibility to free it.
In such case, a manual intervention is required, the builder needs to be
killed and then copr-backend can spawn a new one.

When timeouting the `ansible-playbook' command, we can ultimately kill
the spawning process, making a useless builder instance. However, we
automatize killing such builders via Cron task, so we eventually free
it and get a new one.

Merges: #1091