docker: fix already running httpd issue on distgit
I was trying to hackishly fix this in 6ae31277, but it probably won't
help, because `pidof httpd` is none.

    [root@distgit /]# pidof httpd
    [root@distgit /]# /usr/sbin/httpd -DFOREGROUND
    httpd (pid 35) already running

The mentioned PID is of the following process

    /usr/libexec/git-core/git-daemon --base-path=/var/lib/dist-git/git \
        --export-all --user-path=public_git --syslog --inetd --verbose

Which is started by our supervisord. The question is, do we really need
it? I don't think so, since we can just remove it and the docker compose
environment works just fine.

Merges: #1098