backend: build reattaching: avoid doubled fc.update() call
For each reattached build we called mark_running twice, this both caused
doubled 'starting build: %s' log info message and calls to frontend via
fc.update() (the second one was just wasting of traffic).

Move the 'self.reattach' condition to _announce_start so it is easier to
understand and always call _announce_start() ASAP in Worker.run() (no
matter we reattach or not).

While we are on it, rename mark_running to internal _mark_running.

Merges: #1116