backend: do not start a build if copr_base is not available yet
We don't have any mechanism to define relationships between builds
and actions, so we can possibly come across a situation when an action
for creating repodata for new project takes a longer time than
expected and meanwhile a build in the project is started. That will
inevitably result in failure because we set `skip_if_unavailable=0`
for the `[copr_base]` repository.

Merges: #1125
Fixes: #1057