frontend, doc: put cost=1100 to multilib repo
I noticed that we miss the cost= argument in multilib repofile when
documenting the recently added multilib option.  The should help users
so they don't install the packages from 32/64bit repositories by random.

The docs originally mentioned cost 900/1000, but since the default is
1000 - the main repo should stay default (as with non-multilib
repofiles) and we should special case the multilib repository.  That's
why it has higher cost 1100 for now.