backend: more deterministic test_delete_not_finished_workers
Previously we relied on the fact that background action processor is
able to start (== fork() and communicate with redis) within a very
limited time period, and it often wasn't enough.

Instead of increasing the period, make sure that we control the
time.time() call in worker-manager so we 100% know how many loops happen
in WorkerManager.run().  This requires us to limit the number of calls
to time.time().