backend: copr-repo: createrepo_c --outputdir must imply --baseurl
We could specify "--basedir '../'" instead, but end-user dnf wouldn't
know how to map the relative path to http:// locations (e.g.
http://exmaple.com/results/user/project/chroot/devel/../some.rpm).

So copy the same "hack" from createrepo() method, and specify --baseurl
instead.  This required me to parse backend configuration again, so I
reworked the logic so we do this on one place.

Also we need to guess the ownername/dirname/chroot (same as we needed it
in add_appdata()).  So do this more carefully, and on one place only.

Fixes a bug in --devel detected by beaker tests.