frontend: don't show builds table when there are none of them
This is a regression. Now `builds` variable doesn't contain list of
builds, but rather a SQL query to get them. Therefore condition
{% if builds %} is always true.

We are operating with large data here, so we want to avoid additional
query evaluating to find out if there are any results. The most
performance friendly solution should be for-else. Altough it works
fine, this change is a bit hacky from my point of view.

1. We need to loop the whole table because in the `else` block, we
   don't want to have any table rendered
2. Of course it doesn't make any sense to render `table` and `thead`,
   etc for every build. Just conditionally for `loop.first` and table
   closing when `loop.last`
3. We need to call `build_states()` from within the `builds_table()`.
   Because we want to show states only if there were any builds printed
   and we don't have this information outside of the `build_table()`
   macro.
4. We need to move the `no_builds_window()` code to `_builds_table.html`
   so it can be rendered from the `for-else` else block.

Merges: #1215