frontend: change panel order in 'Rebuild all packages' page
This is a simple panel reordering. When you hit "Rebuild all", it makes
little sense to have the list of packages in the first panel, and
especially in a big project, because you have to scroll down a lot to
see the additional options. So I moved the additional options to the
top.

Merges: #1227