rpmbuild: properly cleanup after obtaining sources
We currently don't clean a temporary directory used for obtaining
sources, so we unintentionally spam /tmp. We don't clean even
when the build succeeds, so this is not related to catching exceptions.

The problem seems to be, that we are trying to clean in `__exit__`
method, which is a mistake. The `__exit__` method executes when
leaving a context - we don't use `*Provider` as a context.
We should clean in destructor instead, hence `__del__`.

Relates: #1245