rpmbuild: fix tests for EPEL6 an 7
The number of `rmtree` calls required for removing a nested directory
differs between Fedora/python3 and EPEL/python2. It is either bug
in `mock` module (which is used only for python2) or the `rmtree`
implementation changed from having recursive calls to having a loop.

I am modifying the test to work on both Fedora and EPEL chroots and
also adding a special case for EPEL to check that those multiple
`rmtree` calls affect only the one desired directory.

Relates: #1262