rpmbuild: don't delete the old .rpmnew files
It is actually blocking our temporary overwrites of mock-core-configs
from koji builds [1] where  we 1) run the script, 2) replace
mock-core-configs with custom version, and 3) run the script again.
When the '.rpmnew' files exist initially, the configs are (likely)
changed and we want them to be replaced.

There's potential that the .rpmnew files are too old (when they are
baked into builder image).  Such situation needs to be worked-around by
explicit removal of those files in ansible.git.

[1] https://infrastructure.fedoraproject.org/cgit/ansible.git/commit/?id=57232452b086c171aa6e1ad1f441acffe7d376df