frontend: not finished build_chroots to disallow copr_chroot removal
When user disables some chroot in his project -> CoprChroot instance is
deleted from DB -> foreign keys from BuildChroot to CoprChroot are
automatically cleared (ondelete set null).

If the BuildChroot was not yet finished, but the corresponding
CoprChroot deleted we wouldn't be able to process the BuildChroot
anymore.  This is perhaps something to be solved in future (BuildChroot
should be self-standing as long as CoprChroot items are removable).

So for now at least guard against inconsistencies in code, if it's not
easily possible directly in DB.

Fixes: #1269
Relates: #1321