frontend: re-assign BuildChroots to re-enabled CoprChroot
If user disables and then re-enables the some chroot in his project,
re-assign the old existing BuildChroot instances to the newly created
CoprChroot instance.

This costs some (I believe trivial) DB work and it isn't absolutely
necessary step.  But it somehow looks reasonable to me at this point in
time.  At least as long as we disallow multiple CoprChroot instances
with the same assigned MockChroot in one Copr project (we could want
this in future, so we would probably revert this patch).

Merges: #1356