frontend: fix repo generation for modules with dash in their name
Fix RHBZ 1833010

The original NSV parsing was too naive and worked only for the most
basic strings, e.g. foo-master-123. When module name contained
a dash, it crashed into error with 500 status code.

I am fixing the parser so it works for such modules and also make
sure that it throws a user-friendly error when the requested NSV
is not valid.