frontend, backend: WorkerManager for builds
The dispatcher logic is moved to Dispatcher class in dispatcher.py, and
both ActionDispatcher and BuildDispatcher inherit from there to re-reuse
as much as possible.  Both dispatchers reference concrete instance of
WorkerManager (RPMBuildWorkerManager or ActionWorkerManager) and now
behave very similarly.

The is_worker_alive() and frontend_client attributes were moved from
ActionWorkerManager to WorkerManager as they are re-usable by
RPMBuildWorkerManager.

The VM allocation logic is obsoleted by resalloc server [1], and instead
of explicitly assigning tasks to VMs we take resalloc tickets and wait
till they are resolved.  Side effect here is that the tasks get a bit
earlier to the build queue (are marked as 'starting') and once the
resalloc ticket is processed they are moved to 'running' state.  The VM
allocation abstraction is in vm_alloc.py, and it better encapsulates the
instance so we can forget about the VM factory back-reference.

The copr-backend-build.service file is synced with
copr-backend-action.service file because we use the same
"BackgroundWorker" concept for builds now.

Because FrontendClient automatically adds trailing slash to each
copr-frontend request, two frontend routes are now fixed so they accept
requests both with/out the trailing slash.

The "reattach" logic is made useless and is removed;  the workers now
continue to run on background even if we restart the copr-backend
service.

Some tests which were skipped before, and now are made irrelevant, were
now dropped.