frontend: more obvious links to live logs
Provide both the backend.log and the builder-live.log links for both
the source and RPM builds.

Provide also the "source state" icon.

Stop using misleading "SRPM" term when we should use only "source" term
instead.  We actually build SRPM, yes, but only to ease importing of the
software into dist-git (and SRPM is actually not needed at all for this
task, this is to be fixed in future) and eventually we anyway build the
"real" SRPM on the target builder (the same builder that builds the
binary RPM, from Copr dist-git).

We also got several complaints before that the links to live-logs are
not visible.  So provide a new column for it in build's results table.