backend: dump attempt to send message to backend.log
There are ATM troubles to send messages in dev environment, so the
(unsuccessful) attempts to send messages take quite some time.  It is
better to dump the info to log so user knows what he is waiting for.

Also, setup the message sender as soon as possible in background worker,
but not in constructor.  The magic in BackgroundWorker.__init__()
replaces the logger object in the meantime, and all the messaging
logging output is lost in stderr only.

Also drop the "decorator" for messaging logger.  The output was not
particularly nice and the additional context isn't needed anyway.