backend: use FileHandler for backend.log
WatchedFileHandler isn't appropriate choice, that's not what we actually
need.  But more importantly, it even caused problems on build cancel:

1. user requested build cancel
2. user requested build delete very quickly
3. the build delete action was processed first, and build was removed
   (including the backend.log)
4. cancel request is processed
5. background worker tracebacks on the first attempt to log into the
   backend.log file because we don't only remove the log file, but whole
   directory.

Where WatchedFileHandler failed before to re-create the file, we simply
ignore the unlink()ed file problem and keep writing into the opened file
descriptor.

As a consequence, because of unexpected error in logging, we did not
even release the allocated VM before (so tickets were kept open, and
resources UP, and unused).