frontend: show _all_ build checkboxes by javascript
This complements b866853e2286cbb8f0d58c3e2e0eb772336da97e.  The original
commit caused a problem that at the time when the whole page document
was loaded - the script responsible for hiding (removing) the javascript
table nodes were already executed.  So the hack

    `$(".show-me-javascript").removeClass("hidden")`

.. only affected a subset of targeted elements (only those which were
currently rendered).

Then, hitting the pagination javascript buttons caused that other nodes
appeared again, but with the (not removed) "hidden" class (so they were
rendered, but invisible).

So now, apply the same hack _before_ we start removing the html nodes by
javascript so all nodes are affected.  Also keep the hack at the bottom
of the page (we may use this idiom on other places not really depending
on the javascript table).

Resolves: rhbz#1846037