rpmbuild: define mock 'root' config for make srpm
When 'root' is _not_ redefined, the caches (and other directories) are
shared with the included default.cfg config file.

During the recent testing of Red Hat Internal Copr instance, it happened
several times in row that heavily re-used VM cached the root
(root-cache.tar.gz) of default.cfg (symlink fedora-32-x86_64.cfg), and
later re-used the cache for make_srpm method.  This failed because when
restoring from the cache the chroot_additional_packages config option
wasn't processed and the reqested dnf package was missing to build the
sources.