rpmbuild: provide the "dynamic" %buildtag
The %buildtag macro can be used similarly to %dist, as post-release
flag.  While the %dist tag is specified by the mock build-chroot,
%buildtag is specified by the wrapping build system.
Example of use is:

    Release: 0%{?dist}%{?buildtag}

This could solve the "release autobump" problem.  The use of this macro
is not at the time of committing this mandatory, nor suggested.  This is
mostly for R&D to see how it serves its purpose.

This commit is reaction on discussion on Fedora devel:
https://lists.fedoraproject.org/archives/list/\
devel@lists.fedoraproject.org/thread/TXLF22CSUVUIQBVHH2NEFF4IOIFHS5WK/

When the %buildtag is in use, simple re-build (resubmit button in
copr UI) of the package causes that the Release is always bumped.

Except that dnf/yum should newly always prefer the latest build of the
affected package in such situation, it also guarantees that our build
garbage collector (prunerepo) really keeps only the latest available
builds of each package (so far, when there are two exactly the same
NVRs, it is not defined which one of those is to be deleted).

Fixes: #840