docker: add resalloc container
Our docker containers doesn't reflect the last rework of copr-backend
code. The copr-backend services are not supposed to spawn builders
anymore and therefore backend-vmm container is not needed anymore.
Instead, we rely on resalloc, for which I am adding a new container
now. Additionally, dispatcher code was reworked and has a single
endpoint now.