python: show obsolete warning on all APIv1 and APIv2 pages
We are already showing a warning about APIv1 and APIv2 being obsolete
but only on their main pages. When somebody goes to the documentation
through google or bookmarks, they might not even notice, that they
are obsolete and that we already use APIv3. Let's just show the
warning on all subpages.

Also some removed whitespace ...