frontend: update the build state descriptions
Some are just revamped, and some are updated to the recent state of
things;  we now (again, after some period of time) start backend workers
before we have a corresponding builder machine in hand.  So the builds
can (again) spend some time in "starting" state, waiting for VMs.  And
it is a bit more correct to differentiate "(Backend) Workers" and
"Builders".