frontend: add EOL repositories page for user (in opposite to project)
Email notifications proved to be not reliable enough and we shouldn't
rely on them. It can happen that emails are not being sent [1], that
they are being delivered as spam [2], or that they are not being
delivered to some users at all [3].

We don't want to create a notification system, that would duplicate
all messages and show them directly at Copr site (at least for the
time being), so we decided to create a page showing outdated
repositories from all projects permissible by a user (in opposite to
having them in project settings). This way, there is a single entry
point for all of them. Users that have problems with receiving emails
from Copr can then bookmark this page and take over the responsibility
and schedule to prolong the duration of their EOL repositories.

[1] https://pagure.io/fedora-infrastructure/issue/9233
[2] https://pagure.io/fedora-infrastructure/issue/9250
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1868367