build_aux: make linter produce diff only for project subdir
Noticed with 'behave' subdirectory;  the "projectdir" is detected as
gitroot (because there's no spec file), and so we previously generated
list of changed files for the whole project (starting from gitroot), not
just behave subdirectory.  Newly, we detect the subdirectory below
"projectdir", check only the list of files related to _that_
subdirectory.

This fix is not only useful for behave subdirectory;  we can as well go
to e.g.  frontend/coprs_frontend/views and hit
../../../build_aux/linter -- and produce issues for _that_ subdirectory
(not for whole frontend).  This may actually speedup the PyLint linter a
lot.