rpmbuild: disable source fetch for the dist-git method
When creating source RPMs from a dist-git repository, we are 100% sure
we don't have to try to download the patches/sources from the Internet
(dist-git repo is self-standing).  So let's disable the automatic
downloads explicitly (no matter if the default is ON/OFF on the host).

We still have the auto-downloading by default ON on copr builder hosts,
but that's just for the purpose of "upload spec" feature, and it is
likely to be changed with #1427.

Fixes: #1219